requirements.txt
asgiref==3.8.1
boto3==1.34.147
botocore==1.34.147
crispy-bootstrap5==2024.2
dj-database-url==0.5.0
Django==4.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-storages==1.14.4
django-summernote==0.8.20.0
django-tinymce==4.0.0
gunicorn==20.1.0
jmespath==1.0.1
pillow==10.2.0
psycopg2==2.9.9
s3transfer==0.10.2
sqlparse==0.4.4
urllib3==1.26.19
whitenoise==5.3.0
python-dotenv>=1.0.1